5 flask frontend (#19)
* flask frontend

* tests dependencies

* read user endpoints

* read user endpoints

* read user endpoints

* user admin ui

* ui delete user

* cleaning

+++ backend/app/api/routes/user.py
@@ -1,9 +1,9 @@
 from fastapi import APIRouter, Depends, HTTPException, status
 from sqlmodel import select
 
-from app.api.crud.user import create_user, delete_user, get_user
-from app.api.deps import SessionDep, is_admin_user
-from app.models.user import UserCreate, UserGroup, UserPublic
+from app.api.crud.user import create_user, delete_user, read_all_users, read_user
+from app.api.deps import CurrentUser, SessionDep, is_admin_user
+from app.models.user import User, UserCreate, UserGroup, UserPublic
 
 router = APIRouter(prefix="/user", tags=["user"])
 
@@ -35,12 +35,12 @@ async def create_user_api(*, user: UserCreate, session: SessionDep) -> UserPubli
     # ensure that no user with the same name or email exists
     # check whether the user already exists
     # TODO should be enforced at the database level and error should be handled
-    if get_user(username=user.username, session=session):
+    if read_user(username=user.username, session=session):
         raise HTTPException(
             status_code=status.HTTP_409_CONFLICT,
             detail="User with the same name already exists",
         )
-    if get_user(username=user.email, session=session):
+    if read_user(username=user.email, session=session):
         raise HTTPException(
             status_code=status.HTTP_409_CONFLICT,
             detail="User with the same email already exists",
@@ -73,6 +73,7 @@ async def create_user_api(*, user: UserCreate, session: SessionDep) -> UserPubli
 async def delete_user_api(
     user_id: int,
     session: SessionDep,
+    current_user: CurrentUser,
 ) -> None:
     """Delete a user. Only admin users can delete users.
 
@@ -80,4 +81,60 @@ async def delete_user_api(
         user_id(int): Identifier of the user
         session (SessionDep): Database session.
     """
+    if current_user.id == user_id and current_user.is_superuser:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Superuser cannot delete itself",
+        )
     delete_user(user_id=user_id, session=session)
+
+
+@router.get("/me", response_model=User)
+async def read_user_me(current_user: CurrentUser) -> User:
+    """
+    Get current user.
+
+    Returns:
+        User: Current user data.
+    """
+    return current_user
+
+
+@router.get("/{user_id}", response_model=User)
+async def read_user_api(
+    user_id: int, current_user: CurrentUser, session: SessionDep
+) -> User:
+    """
+    Get user by id
+
+    Args:
+        user_id (int | str): User id or mail address.
+
+    Returns:
+        User: User data.
+    """
+    if user_id == current_user.id:
+        return current_user
+    if current_user.is_superuser:
+        user = read_user(user_id=user_id, session=session)
+        if user:
+            return user
+        else:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail="User not found",
+            )
+    raise HTTPException(
+        status_code=status.HTTP_403_FORBIDDEN,
+        detail="Not enough permissions",
+    )
+
+
+@router.get(
+    "/",
+    response_model=list[User],
+    status_code=status.HTTP_200_OK,
+    dependencies=[Depends(is_admin_user)],
+)
+async def read_users(session: SessionDep) -> list[User]:
+    return read_all_users(session=session)
